+++ src/gamescene/game.cpp
@@ -675,7 +675,7 @@ void game_scene::HANDLE_UPDATE(player_hp, const player_hp_update &args) {
     int prev_hp = player->hp;
     player->hp = args.hp;
     if (args.instant) {
-        player->set_hp_marker_position(args.hp);
+        player->set_hp_marker_position(static_cast<float>(args.hp));
     } else if (prev_hp != args.hp) {
         add_animation<player_hp_animation>(options.move_hp_msecs, player, prev_hp);
     }
@@ -692,7 +692,7 @@ void game_scene::HANDLE_UPDATE(player_show_role, const player_show_role_update &
             p->m_role->make_texture_front(parent->get_renderer());
             if (args.instant) {
                 if (args.role == player_role::sheriff) {
-                    p->set_hp_marker_position(++p->hp);
+                    p->set_hp_marker_position(static_cast<float>(++p->hp));
                 }
                 p->m_role->flip_amt = 1.f;
             } else {
